Added Leg Tracking
Added in new character model with leg tracking tied to the rotation of 2 vive trackers mounted on the top of the shoes of the player. Will implement capability for this tracking to done at the knee / thigh as well. Cleaned up blueprint to make it readable and neater as well.